requirements.txt
# IT Audit Team Skills — Python Dependencies
# Install with: pip install -r requirements.txt

# Cloud integrations
boto3>=1.26.0
botocore>=1.29.0
msal>=1.20.0

# Environment variable management
python-dotenv>=1.0.0

# Data handling
pandas>=1.5.0
tabulate>=0.9.0
pyyaml>=6.0

# Templating
jinja2>=3.1.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Code quality
flake8>=6.0.0
mypy>=1.0.0

# Type stubs for mypy
boto3-stubs[iam,cloudtrail,ec2,s3]>=1.26.0
types-tabulate>=0.9.0
types-pyyaml>=6.0.0
